Linux buildfixes. Use SSE by default instead of AVX.

+++ network.cpp
@@ -8,15 +8,15 @@ void NeuralNetwork::ClearDeltaWeightSum() {
 }
 
 // Inference.
-void NeuralNetwork::RunForwardPass(int miniBatchSize) {
+void NeuralNetwork::RunForwardPass(size_t miniBatchSize) {
 	for (int i = 0; i < layers.size(); i++) {
-		layers[i]->Forward(miniBatchSize, i > 0 ? layers[i - 1]->data : nullptr);
+		layers[i]->Forward((int)miniBatchSize, i > 0 ? layers[i - 1]->data : nullptr);
 	}
 }
 
-void NeuralNetwork::RunBackwardPass(int miniBatchSize) {
+void NeuralNetwork::RunBackwardPass(size_t miniBatchSize) {
 	for (int i = (int)layers.size() - 1; i >= 0; i--) {
-		layers[i]->Backward(miniBatchSize,
+		layers[i]->Backward((int)miniBatchSize,
 			i > 0 ? layers[i - 1]->data : 0,
 			(i < layers.size() - 1) ? layers[i + 1]->gradient : nullptr);
 	}
